+++ src/noshow/preprocessing/load_data.py
@@ -5,7 +5,7 @@ from typing import Dict, List, Union
 import pandas as pd
 
 from noshow.api.pydantic_models import Appointment
-from noshow.config import ClinicConfig
+from noshow.config import NO_SHOW_CODES, ClinicConfig
 
 
 def load_appointment_pydantic(input: List[Appointment]) -> pd.DataFrame:
@@ -91,11 +91,9 @@ def process_appointments(
     """
     appointments_df = apply_config_filters(appointments_df, clinic_config, start_date)
 
-    no_show_codes = ["N", "0000000012", "0000000013"]
-
     appointments_df["no_show"] = "show"
     appointments_df.loc[
-        appointments_df["cancelationReason_code"].isin(no_show_codes), "no_show"
+        appointments_df["cancelationReason_code"].isin(NO_SHOW_CODES), "no_show"
     ] = "no_show"
 
     # Some patients have multiple postal codes